Add styles and restructure button component
